Split Query into a separate file
Keeping it separated from Forge for now
